rename logging file to report and remove old report file
